(PE-31356) Add individual creds fields
This commit adds the `client_email`, `token_uri` and `private_key` fields
to the plugin to allow users to pass in the relevant credentials directly
rather than using a file. This is to avoid relying on external
files/variables for the task. If a credentials file is provided, it will
take precedence over these new credential parameters.